fix(contributions): remove `when` prop from CommandContribution
The `when` field should live on the type that associates a menu with a command, because you may want the command to be shown in some menus and not others. It is not necessary to have two `when`s.

BREAKING CHANGE: The `CommandContribution#when` field is removed. Use `MenuItemContribution#when` instead.