feat: new extension API that only requires extensions to export 1 func
This makes it so that extension JS bundles are expected to merely export a single activate function. They are no longer bundled with the sourcegraph module, and they no longer know how to speak the RPC extension protocol themselves. The "extension host" knows how to invoke their activate function and inject the sourcegraph module (when it is imported).

The goal is to make it easier to write extensions. You just need to write them against a TypeScript API. Your extension itself is not responsible anymore for executing itself, communicating with the client, or knowing anything about its JavaScript execution environment.

For example, the following is now a valid extension:

```
import * as sourcegraph from 'sourcegraph'

export function activate(): void {
    sourcegraph.registerHoverProvider(['*'], {
        provideHover: () => ({
            contents: { value: ' Hello from an extension!' },
        }),
    })
}
```

When bundled, it will contain ONLY that `activate` function; the `sourcegraph` module is dynamically linked at runtime. This means the extension bundle is significantly smaller, and the extension author only needs to care about adhering to the TypeScript API (not about also bundling their extension with the latest version of the `sourcegraph` package for wire protocol compatibility).